Refactor: Moved DocumentInputs
Need this to be in the "core" project so that the langchain4j-specific embedder can access it for batching purposes.

Expecting DocBuilder to eventually go into the core projet as well.
