+++ marklogic-spark-connector/src/main/java/com/marklogic/spark/writer/file/ArchiveFileIterator.java
@@ -7,9 +7,9 @@ import com.marklogic.client.io.DocumentMetadataHandle;
 import com.marklogic.client.io.Format;
 import com.marklogic.client.io.InputStreamHandle;
 import com.marklogic.spark.Util;
+import com.marklogic.spark.core.DocumentInputs;
 import com.marklogic.spark.reader.document.DocumentRowSchema;
 import com.marklogic.spark.reader.file.ArchiveFileReader;
-import com.marklogic.spark.writer.DocBuilder;
 import org.apache.commons.io.IOUtils;
 import org.apache.spark.sql.catalyst.InternalRow;
 
@@ -21,7 +21,7 @@ import java.util.Iterator;
  * {@code DocumentRowConverter} to build sets of document inputs from an archive file without reading any content entry
  * into memory - thus supporting streaming of an archive.
  */
-public class ArchiveFileIterator implements Iterator<DocBuilder.DocumentInputs>, Closeable {
+public class ArchiveFileIterator implements Iterator<DocumentInputs>, Closeable {
 
     private final ArchiveFileReader archiveFileReader;
     private final Format documentFormat;
@@ -40,7 +40,7 @@ public class ArchiveFileIterator implements Iterator<DocBuilder.DocumentInputs>,
     // Suppressing sonar warning about throwing a NoSuchElementException. We know this is only used by
     // DocumentRowConverter, which properly calls hasNext() before calling next().
     @SuppressWarnings("java:S2272")
-    public DocBuilder.DocumentInputs next() {
+    public DocumentInputs next() {
         InternalRow row = archiveFileReader.get();
         String uri = row.getString(0);
         if (Util.MAIN_LOGGER.isDebugEnabled()) {
@@ -51,7 +51,7 @@ public class ArchiveFileIterator implements Iterator<DocBuilder.DocumentInputs>,
         if (this.documentFormat != null) {
             contentHandle.withFormat(this.documentFormat);
         }
-        return new DocBuilder.DocumentInputs(uri, contentHandle, null, metadata);
+        return new DocumentInputs(uri, contentHandle, null, metadata);
     }
 
     @Override
