+++ marklogic-spark-connector/src/main/java/com/marklogic/spark/writer/file/FileIterator.java
@@ -4,7 +4,7 @@
 package com.marklogic.spark.writer.file;
 
 import com.marklogic.client.io.InputStreamHandle;
-import com.marklogic.spark.writer.DocBuilder;
+import com.marklogic.spark.core.DocumentInputs;
 import org.apache.commons.io.IOUtils;
 
 import java.io.Closeable;
@@ -15,12 +15,12 @@ import java.util.stream.Stream;
  * Exists solely to provide an implementation of {@code Closeable} so that the {@code InputStreamHandle} can be closed
  * after the corresponding document is written to MarkLogic.
  */
-public class FileIterator implements Iterator<DocBuilder.DocumentInputs>, Closeable {
+public class FileIterator implements Iterator<DocumentInputs>, Closeable {
 
     private final InputStreamHandle contentHandle;
-    private final Iterator<DocBuilder.DocumentInputs> iterator;
+    private final Iterator<DocumentInputs> iterator;
 
-    public FileIterator(InputStreamHandle contentHandle, DocBuilder.DocumentInputs inputs) {
+    public FileIterator(InputStreamHandle contentHandle, DocumentInputs inputs) {
         this.contentHandle = contentHandle;
         this.iterator = Stream.of(inputs).iterator();
     }
@@ -31,7 +31,7 @@ public class FileIterator implements Iterator<DocBuilder.DocumentInputs>, Closea
     }
 
     @Override
-    public DocBuilder.DocumentInputs next() {
+    public DocumentInputs next() {
         return this.iterator.next();
     }
