+++ marklogic-spark-connector/src/main/java/com/marklogic/spark/writer/file/GzipFileIterator.java
@@ -5,8 +5,8 @@ package com.marklogic.spark.writer.file;
 
 import com.marklogic.client.io.Format;
 import com.marklogic.client.io.InputStreamHandle;
+import com.marklogic.spark.core.DocumentInputs;
 import com.marklogic.spark.reader.file.GzipFileReader;
-import com.marklogic.spark.writer.DocBuilder;
 import org.apache.commons.io.IOUtils;
 
 import java.io.Closeable;
@@ -17,10 +17,10 @@ import java.util.stream.Stream;
  * Exists solely to provide an implementation of {@code Closeable} so that the {@code GzipFileReader} can be closed
  * after the corresponding document is written to MarkLogic.
  */
-public class GzipFileIterator implements Iterator<DocBuilder.DocumentInputs>, Closeable {
+public class GzipFileIterator implements Iterator<DocumentInputs>, Closeable {
 
     private final GzipFileReader gzipFileReader;
-    private Iterator<DocBuilder.DocumentInputs> iterator;
+    private Iterator<DocumentInputs> iterator;
 
     public GzipFileIterator(GzipFileReader reader, Format documentFormat) {
         this.gzipFileReader = reader;
@@ -30,7 +30,7 @@ public class GzipFileIterator implements Iterator<DocBuilder.DocumentInputs>, Cl
         if (documentFormat != null) {
             contentHandle.withFormat(documentFormat);
         }
-        this.iterator = Stream.of(new DocBuilder.DocumentInputs(uri, contentHandle, null, null)).iterator();
+        this.iterator = Stream.of(new DocumentInputs(uri, contentHandle, null, null)).iterator();
     }
 
     @Override
@@ -39,7 +39,7 @@ public class GzipFileIterator implements Iterator<DocBuilder.DocumentInputs>, Cl
     }
 
     @Override
-    public DocBuilder.DocumentInputs next() {
+    public DocumentInputs next() {
         return this.iterator.next();
     }
