+++ marklogic-spark-connector/src/main/java/com/marklogic/spark/writer/rdf/TriplesDocument.java
@@ -4,7 +4,7 @@
 package com.marklogic.spark.writer.rdf;
 
 import com.marklogic.client.extra.jdom.JDOMHandle;
-import com.marklogic.spark.writer.DocBuilder;
+import com.marklogic.spark.core.DocumentInputs;
 import org.apache.spark.sql.catalyst.InternalRow;
 import org.jdom2.Document;
 import org.jdom2.Element;
@@ -52,9 +52,9 @@ class TriplesDocument {
         return tripleCount >= TRIPLES_PER_DOCUMENT;
     }
 
-    DocBuilder.DocumentInputs buildDocument() {
+    DocumentInputs buildDocument() {
         JDOMHandle content = new JDOMHandle(document);
         String uri = String.format("/triplestore/%s.xml", UUID.randomUUID());
-        return new DocBuilder.DocumentInputs(uri, content, null, null, graph);
+        return new DocumentInputs(uri, content, null, null, graph);
     }
 }
